backend/requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.0
geopy==2.4.1
pydantic==2.9.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
python-multipart==0.0.9

